Refactor Jumping Fountain paint code to paint/entity
