+++ src/openrct2/SpriteIds.h
@@ -874,6 +874,9 @@ enum : ImageIndex
     SPR_EXPLOSION_FLARE = 22896,
     SPR_CRASH_SPLASH_PARTICLE = 22927,
 
+    SPR_JUMPING_FOUNTAIN_WATER = 22973,
+    SPR_JUMPING_FOUNTAIN_SNOW = 23037,
+
     SPR_LAND_OWNERSHIP_AVAILABLE = 22955,
     SPR_LAND_CONSTRUCTION_RIGHTS_AVAILABLE = 22956,
