+++ src/openrct2/entity/Fountain.h
@@ -13,8 +13,6 @@
 #include "../world/Location.hpp"
 #include "EntityBase.h"
 
-struct PaintSession;
-
 namespace OpenRCT2
 {
     class DataSerialiser;
@@ -52,7 +50,6 @@ namespace OpenRCT2
         void Update();
         static void StartAnimation(JumpingFountainType newType, const CoordsXY& newLoc, const TileElement* tileElement);
         void Serialise(DataSerialiser& stream);
-        void Paint(PaintSession& session, int32_t imageDirection) const;
 
     private:
         JumpingFountainType GetType() const;
