+++ src/openrct2/libopenrct2.vcxproj
@@ -400,6 +400,7 @@
     <ClInclude Include="paint\Painter.h" />
     <ClInclude Include="paint\Balloon.h" />
     <ClInclude Include="paint\entity\Paint.Duck.h" />
+    <ClInclude Include="paint\entity\Paint.Fountain.h" />
     <ClInclude Include="paint\entity\Paint.Guest.h" />
     <ClInclude Include="paint\entity\Paint.Litter.h" />
     <ClInclude Include="paint\entity\Paint.Peep.h" />
@@ -1000,6 +1001,7 @@
     <ClCompile Include="paint\PaintHelpers.cpp" />
     <ClCompile Include="paint\entity\Paint.Balloon.cpp" />
     <ClCompile Include="paint\entity\Paint.Duck.cpp" />
+    <ClCompile Include="paint\entity\Paint.Fountain.cpp" />
     <ClCompile Include="paint\entity\Paint.Guest.cpp" />
     <ClCompile Include="paint\entity\Paint.Litter.cpp" />
     <ClCompile Include="paint\entity\Paint.Peep.cpp" />
