+++ src/openrct2/paint/entity/Paint.Fountain.h
@@ -1,0 +1,21 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2026 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#pragma once
+
+#include <cstdint>
+
+namespace OpenRCT2
+{
+    struct JumpingFountain;
+}
+
+struct PaintSession;
+
+void PaintFountain(PaintSession& session, const OpenRCT2::JumpingFountain& fountain, int32_t imageDirection);
